flow fields, kernels, yadda

+++ engine/lib/object.dart
@@ -3,8 +3,8 @@ import 'package:crawlspace_engine/shop.dart';
 
 class SpaceObject {
   final String name;
-  int fedLvl;
-  int techLvl;
+  double fedLvl;
+  double techLvl;
   bool known = false;
   String description = "";
   Shop? shop, yard;
